[build] Introduce basic_toolchain()
BUILDCONFIG.gn relies on the fact that toolchain_variant.base is
defined to a non-default value to detect that it is not executed
in the context of the default toolchain.

Unfortunately, several toolchain() definitions did not set
toolchain_variant.base properly, resulting in extra computations
performed in BUILDCONFIG.gn when they were not necessary.

This CL does the following:

- Introduces a new template, basic_toolchain(), that can
  be used to define a toolchain() instance that only uses
  action() targets, while ensuring that toolchain_variant.base
  is always set properly.

  Additionally, the template asserts that the toolchain's
  label correspond to the right expected value, just in case.

- Update non-Clang based toolchain() definitions to use
  basic_toolchain().

- Move the logic to detect the default toolchain to the
  in_default_toolchain variable definition in BUILDCONFIG.gn,
  and document why it is here.

For testing, I manually verified that the 'gn gen' output was
the same before and after the CL is applied, except for
out/default/build.ninja.d which now lists the new
build/toolchain/basic_toolchain.gni as input dependency.

Bug: None
Testing: See above.
Change-Id: I8ff573166463b6a1ae2b2673c9f1c25007550a6d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/417836
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>