[storage] Explicitly zero fill in compressed FVM sparse images.
At the moment in FVM sparse images, there's no way to tell apart cases
where zero-fill is required and where we don't care. Since zeroes
compress really well, this change is explicit about zero filling and
indicates via a flag in the header that zero filling is then not
required. I've tested that this makes a neglible different to generated
size.

This should help a small amount with flash wear on devices that do a lot
of paving/flashing because we don't need to zero fill so much.

Change-Id: Ib2d0afd64cda729ad6b7444651e018092c4e361f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/419536
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>
Testability-Review: Gianfranco Valentino <gevalentino@google.com>