[camera][imx355] Fix crash on boot
This CL fixes a crash when booting a device that uses the imx355
driver. The root cause was the ISP query of the number of
available modes, which asked for more sensor modes than were available.

The fix was to change the logic to return the smaller of the
number of modes available and the number that was requested.

This CL also removes some spurious driver logging during driver
initialization.

Tested on a luis platform board with the camera in both the
muted and unmuted states at power on.

Change-Id: I834875399b5fa0dd9ef9f9e08a03ff1a464aa2d9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/419754
Testability-Review: Nic Zoghb <nzo@google.com>
Commit-Queue: John Sasinowski <jsasinowski@google.com>
Reviewed-by: Ed James <edjames@google.com>
Reviewed-by: Carl Norum <cjn@google.com>