[audio][amlogic] Make sure codecs are muted when starting
This CL introduces an override_mute flag that is always checked
when updating the gain state in the codec(s). With this we can
guarantee that we'll always set the mute flag in the codecs before
we actually start a playback. Added tests to confirm we always
set muted in the codecs upon boot, and also to check the logic
on updating the gain state (including the added aggregation of
codecs mute and agc reporting).

Test: aml-g12-tdm-test
Test: Manual check of YouTube audio in Astro and Sherlock
Test: No pop on reboot on Astro and Sherlock
Change-Id: Ice045546fb230c86c12ebf39adc52ef668f1f45a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/418234
Reviewed-by: Braden Kell <bradenkell@google.com>
Reviewed-by: Eric Holland <hollande@google.com>
Testability-Review: Braden Kell <bradenkell@google.com>
Commit-Queue: Andres Oportus <andresoportus@google.com>