[debugger] Give process name to empty module name.
By conventoion the dynamic loader sets the module name of the main
binary to empty. This produces output from "libs" and "sym-stat" which
is confusing to users.

This patch fills in the process name for empty module names. This isn't
always guaranteed to be perfect, but at least gives some indication of
what the binary is.

Bug: 5233
Change-Id: Ic7df53e5de63a3f684d0d2746c3b8be4fb3684b4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/419922
Reviewed-by: Dangyi Liu <dangyi@google.com>
Testability-Review: Brett Wilson <brettw@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>