[vulkan] Return an error when fence passed to vkAcquireNextImageKHR
The swapchain won't ever signal a fence passed to vkAcquireNextImageKHR,
and it's not possible to implement signaling that fence  - we don't
support external fences to signal it on the CPU, and no queue is passed
in that could be used to signal it from Vulkan.

There's an assert to protect against callers passing in a fence, but the
copy in the SDK is compiled without asserts enabled. So signal a debug
callback and return an error in that case.

Multiply: image-pipe-swapchain-tests

Change-Id: Ic1c5bb074b676207a7556539caa8a5c25fc6f00d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/418955
Commit-Queue: John Bauman <jbauman@google.com>
Testability-Review: John Bauman <jbauman@google.com>
Testability-Review: Craig Stout <cstout@google.com>
Reviewed-by: Craig Stout <cstout@google.com>