[io][conformance] Fix rustvfs test harness so it builds
Fixed up a few issues with the BUILD.gn file, and removed the package
definition for the rustvfs harness (it wasn't formatted correctly, and
isn't actually needed, the rustc_binary is good enough).

Also fixed up the Rust code for the test harness to get it to build. It
was using old versions of the FIDL protocol and VFS structs.

I can't promise it actually works yet, but at least it builds now. Next
I will hook it up to the conformance test runner and fix any issues in
the implementation that crop up.

Bug: 33880
Change-Id: I3d76c6902bd0dedec728df5aa0d9a86c48d67a11
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/420074
Reviewed-by: Chris Suter <csuter@google.com>
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Testability-Review: Chris Suter <csuter@google.com>
Commit-Queue: Cameron Silvestrini <csilvestrini@google.com>