[triage] Separate error string for MISSING
This is a temporary solution until the "missing" refactor is done. For
now it is useful to know if there was a parse error or if something is
just missing.

TESTED=Manual fx triage. This is human-readable output only and should
not be asserted in tests.

Change-Id: I971b07ab4e6351486319a5a9488f8df802439e61
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/420330
Testability-Review: Christopher Johnson <crjohns@google.com>
Commit-Queue: Christopher Johnson <crjohns@google.com>
Reviewed-by: Chris Phoenix <cphoenix@google.com>