[a11y] Screen Reader unit test refactor.
This change refactors the Screen Reader unit test.

The unit test, as is right now, is not testing Screen Reader behavior, but the actions themselves. This makes hard to write new tests for the Screen Reader, because they need to replicate the action infrastructure.

This change makes the Screen Reader test its own behavior, mocking the result of the actions (which are already tested by the actions), and also make an abstraction to call the gestures (which are already tested in other places).

Fixed: 49263
Change-Id: If9a0b802136cd4bf0c069390da072a93d1c66a94
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/419478
Commit-Queue: Lucas Radaelli <lucasradaelli@google.com>
Reviewed-by: Alice Neels <neelsa@google.com>
Testability-Review: Alice Neels <neelsa@google.com>