[kernel][vm] Explict page list node ranges
Modifies the VmPageListNode to have separate ForEveryPage and
ForEveryPageInRange methods, with the latter requiring that it be given
a range that is strictly inside the nodes range.

The VmPageList::ForEveryPageInRange is then modified to iterate the
range and know whether it is operating on a partial node range or a
full node range instead of throwing ranges at nodes and seeing what
sticks.

Motivation for this change is that knowing and being explicitly in the
range being operated on allows for future refactors and improvements.

Test: Covered by existing tests.

Change-Id: Iaa8c55af566384ae876ca4a0a42a95fdabd7bdd1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/416539
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Testability-Review: Rasha Eqbal <rashaeqbal@google.com>