[kernel][vm] Avoid race in user_ids
This removes a race where a VMO could be destroyed before its child
had its user_id set. For copy-on-write VMOs destruction potentially
requires re-attributing all the hidden parents. Re-attributing with
an unset user-id is both a completely incorrect thing to do, and
also trips various assertions. The solution here is to just hold
onto the parent VMO refptr after creating the child until the
user_id is set.

The only way to exploit this I have been able to come up with requires
a total of 4 threads. Assuming we have an initial VMO the first 3
threads all attempt to create children. Now if each of those three
threads gets preempted just prior to calling set_user_id in
VmObjectDispatcher::Create then we will have a VMO tree that looks like

      |A|
      / \
    |A| |0|
    / \
  |A| |0|
  / \
|A| |0|

Here each VMO is labeled with its attribution, with A being the koid of
the dispatcher of the original VMO. The three children have not had
their user_id set yet, and so there attribution is 0.

If we close the original VMO (which is now the bottom left leaf node),
as all the create_child syscalls dropped their refptr the VMO gets
cleaned up. This results in the hidden VMO of its immediate parent
being merged into the sibling, and taking on its siblings attribution.

    |A|
    / \
  |A| |0|
  / \
|0| |0|

Now we need to walk up the tree and perform reattribution, since the
user_id A no longer exists. For the first step the logic of the loop
requires us to simply not pick A, which we do and end up with:

    |A|
    / \
  |0| |0|
  / \
|0| |0|

At this point we recorded user_id_to_skip as the last id we set (which
is 0) and so for the next reattribution we require that we end up
picking !A and !0, except since both children have the attribution of 0,
we have to pick it, and thus the assert fires.

The assumption being invalidated here is that all leaf VMOs have unique
user_ids, which they should since dispatchers should have unique koids.
In this case the unset user ID is acting as a mechanism to have
multiple VMOs share a user id.

Test: The time window of the race is extremely narrow and needs to
be coordinated across 4 different threads. I was unable to come up with
reproductions that did not involve adding Sleep()s in the syscall path.

Bug: 58028

Change-Id: I6f8e88c97eb4ea7c5d48194163d5c58c75f97b96
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/419041
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Testability-Review: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>