[fidl] Add check() to QueryResponseFut
Adds check() to QueryResponseFut, allowing an early check for the
message having been sent.  This is similar to the check that is
performed by one-way messages that have no response.  If an error
occurred in sending, it is returned, otherwise a QueryResponseFut is
returned that can be polled for completion.

Test: added client_query_result_check
      updated other tests for client_raw_query API change

Multiply: fidl-rust-tests

Change-Id: I6ebf8ea5e348717fb35f04f9d88afb185bf9b19f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/418766
Reviewed-by: Mitchell Kember <mkember@google.com>
Reviewed-by: Adam Lesinski <adamlesinski@google.com>
Testability-Review: Mitchell Kember <mkember@google.com>
Testability-Review: Adam Lesinski <adamlesinski@google.com>
Commit-Queue: Marie Janssen 💖 <jamuraa@google.com>