[syscalls][kazoo] Map bool in syscall fidl to uint32_t in C
This is necessary because the compiler is allowed to assume that the
incoming bool is either 0 or 1 on the incoming side, resulting in
incorrect code generation (but with no way for the kernel side to
actually validate the value). By making it an explicit integer type, the
kernel side can range-check.

This causes a change from bool to uint32 in C, Rust, and Go bindings,
however there are currently no bools used in syscall fidl after
https://fuchsia-review.googlesource.com/c/fuchsia/+/418836. As a result
there is no (further) change to the syscall api exposed by the system.

Bug: 53118
Change-Id: I528efe70af7477fdf0c0376de788475cbd4438fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/419682
Reviewed-by: Christopher Anderson <cja@google.com>
Testability-Review: Scott Graham <scottmg@google.com>
Commit-Queue: Scott Graham <scottmg@google.com>