[resultdb] Update logic for setting the "expected" field.
We have tests that fails expectedly (ex:
host_x64/zxtest-integration-test). We are storing the result correctly
in summary.json where the testSuite reports the test as "PASSED", and
the testCase reports tests as PASSED/FAILED based on what really
happened from stdout.

We want to record the "expected" field in ResultDB to indicate the
correct test behavior. If a "FAIL" is correct, then "expected" field
will be set to True; and if a "FAIL" is incorrect (99% of the time), the
"expected" field will be set to False.

Bug: 58744
Change-Id: I349ebb5747b6b6f76f4abc5527e04911335bd911
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/419736
Commit-Queue: Yuan Zhi <yuanzhi@google.com>
Reviewed-by: Gary Miguel <garymm@google.com>
Testability-Review: Gary Miguel <garymm@google.com>