[blobfs] Fix flakiness in AllocateIncrementsMetricTest
The inspect reader service was incorrectly making the first recursive
call to the Tree FIDL client immediately on the calling thread and
successive calls would be wrapped in a promise that would run on the
executor, possibly in another thread.

This is bad behavior because the FIDL client is thread-hostile
and caused flaky behavior every 5000 runs or so.

This CL makes the following changes:
* ensures that inspect::SnapshotTreeFromTree calls are wrapped
  in a promise that will run only on the expected thread.
* in AllocateIncrementsMetricTest, quits the loop and joins on
  the thread before destroying the Executor. This ensures there
  are no running tasks when the Executor gets destroyed.

Bug: 55829
Test: `fx shell runtests -n blobfs-integration-test -- -i 1000`
Change-Id: I35bdf76d00c70d1cdb934e7cd72cb36efdeff1f2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/421114
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Testability-Review: James Sullivan <jfsulliv@google.com>