[ui-input][root_presenter] Add multiple views
In preparation for converting RootPresenter to using the new
fuchsia.ui.pointerinjector API for input injection.
Inserts two extra views internal to RootPresenter to use as context and
target for injection.

Adjusts transform logic to make this easier:
The scene and the root view is now matched to the native rotation and
resolution of the display.
The internal child view (which will eventually be the
"accessibility view" fxb/50549) is rotated and scaled to match the pips
instead.
This allows input to be injected raw and have scenic handle any
necessary transformations.

Bug: 49819
Manually tested on emu, astro and sherlock.

Change-Id: Id1905e7bd6d1224b864a467d9a9219a199fb63fe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407728
Reviewed-by: Jaeheon Yi <jaeheon@google.com>
Testability-Review: Jaeheon Yi <jaeheon@google.com>
Commit-Queue: Mikael Pessa <mikaelpessa@google.com>