[audio][test] Fix min lead time race
In some tests, when starting a renderer, the following event sequence is
rare but possible:

1. Create renderer
2. Ask for MinLeadTime events
3. OnMinLeadTimeChanged(0)
4. Renderer links to output
5. OnMinLeadTimeChanged(>0)

We are using the first OnMinLeadTimeChanged event as a signal that the
renderer has been linked to the output. Per the above sequence, we
should actually wait until the lead time is non-zero.

Fixed: 58394
Change-Id: I21ab1a7b5c7b9a3b4688da5e891bd5dca05f9e73
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/420817
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Testability-Review: Tim Detwiler <tjdetwiler@google.com>
Commit-Queue: Tom Bergan <tombergan@google.com>