[SL4F] Add Reading DDR windowing data
Allows reading a 'sticky' register which stores
DDR windowing tool results that were generated
by the bootloader.

Included in the CL are the changes to the fidl API,
the implementation in the driver, and the facade
which reads the value for factory testing.

Test: Facade tests are included. A follow-up CL
allows manually testing the driver via the CLI.
An integration test of the facade is in a followup CL
in the factory repo.

Change-Id: I0a079ad5b10cd82996d2d061fd76fab8f3c208f9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/419317
Commit-Queue: Garratt Gallagher <garratt@google.com>
Reviewed-by: Braden Kell <bradenkell@google.com>
Reviewed-by: Carlos Pizano <cpu@google.com>
Reviewed-by: Gurjant Kalsi <gkalsi@google.com>
Reviewed-by: Didrik Nordstrom <dnordstrom@google.com>
Testability-Review: Braden Kell <bradenkell@google.com>
API-Review: Ruchira Ravoori <ravoorir@google.com>
API-Review: Carlos Pizano <cpu@google.com>