[build][alln] Improve zx_library() and zx_host_tool() in the Fuchsia build.
This CL provides several improvements to the zx_library() and
zx_host_tool() templates, as defined in the Fuchsia build, in
order to better match the behaviour of the same templates as they
are implemented by the Zircon build.

It also adds the `sdk_migrated` parameter to these templates in the
Zircon build, in order to indicate that a given target should be
built directly with the Fuchsia build. This prevents the Zircon build
from generating a legacy target entry in one of the
out/default.zircon/legacy_[host]_targets-$cpu.json files.

Note that the zx_library() implementation will rewrite deps and
public_deps values to replace references to :headers sub-targets
into regular library targets. This is what the legacy metadata
export mechanism is already doing in the Zircon build.

Note that for non-kernel libraries, it is not possible to create
headers sub-targets, since these conflicts with other features
of the Fuchsia build (e.g. use of 'public', and SDK dependencies).

Similarly, references to //zircon/system/ulib/zircon are replaced
by sysroot references when appropriate.

More precisely, after this CL is submitted, one can ensure that a
legacy target is built directly with the Fuchsia build by doing the
following:

  - Add a symlink from BUILD.gn to BUILD.zircon.gn in the target's
    source directory. This ensures the same high-level definitions
    are used for the target in both builds.

  - Add the following fragment to the start of BUILD.zircon.gn:

      # This BUILD.zircon.gn file is symlinked to BUILD.gn
      if (zx != "/") {
        import("//build/unification/zx_library.gni")
      }

  - Add `sdk_migrated = true` to the target definition,
    i.e. inside a zx_library() or zx_host_tool() call.

  - Replace all occurences of //zircon/public/lib/<name>
    with //zircon/system/ulib/<name> in BUILD.gn files.

A future CL will demonstrate this for the
//zircon/system/ulib/cmdline library (which uses zx_library())
and another one for //zircon/tools/kazoo (which uses zx_host_tool())

+ Add logic to zx_library() to support "kernel = true".
  Properly replicating the logic of the Zircon build requires
  creating ":headers" sub-targets and ":headers.config" configs
  at the moment.

For testing, the output of "gn gen" has been checked manually before
and after this CL was applied. Both for out/default and
out/default.zircon. Since sdk_migrated is not used yet, they are
still identical.

Bug: 54161
Fixed: 57367
Testing: See above.

Change-Id: I60b87795b9311ed67fa727d83ce848bca53a8b30
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/420277
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: P.Y. Laligand <pylaligand@google.com>