[build][alln] Add $zx_build/public/gn/migrated_targets.gni
Add a new .gni file which will be included by the BUILD.zircon.gn
and BUILD.gn files of so-called Zircon legacy targets, which need
to be built both by the Zircon and Fuchsia build, until unification
completes.

In practice, this means that such a target can be migrated to build
properly with the Fuchsia build with a scheme like:

  - Make BUILD.gn a symlink to BUILD.zircon.gn in the same directory
    (to ensure both are kept in sync).

  - Add import("$zx_build/public/gn/migrated_targets.gni") at the
    start of BUILD.zircon.gn (and hence BUILD.gn too). This will
    be a no-op for the Zircon build, but will define zx_library()
    and other templates for the Fuchsia build.

  - Set `sdk_migrated = true` in its zx_library() and zx_host_tool()
    calls.

  - Replace any reference to //zircon/public/lib/<name>/ with
    //zircon/system/ulib/<name> in the rest of BUILD.gn files to
    point to the correct target.

Bug: 58463
Testing: No build change for now.
Change-Id: I8d934ec8f87428a12795037c4d2439463db3086f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/420374
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Commit-Queue: David Turner <digit@google.com>