[kernel][vm] VMO hierarchy generation counts
Add a generation count associated with each VmObjectPaged hierarchy,
which is incremented on any operation that could result in a change in
page attribution counts anywhere in the hierarchy. The generation count
is used to implement caching of page attribution counts (changes for
this to follow in a separate commit).

This commit just introduces the generation count, and increments it in
the required places. The generation count is not used anywhere yet.

These are the spots where the generation count is updated.
  1. Creation and removal of clones and slices.
  2. Committing a new page in GetPageLocked().
  3. Resizing the VMO, decommitting or zero'ing a range.
  4. Taking and supplying pages via supply_pages().
  5. Evicting a page, or deduping a zero page.

Bug: 56779
Change-Id: Ic59e373ee14c3f71297e73faf43d5cffe7919ab0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/417454
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Testability-Review: Adrian Danis <adanis@google.com>