[amlogic_decoder] Heuristic for lower latency decode
Since we don't presently get max_num_reorder_frames
from the FW, for now we can lean on this heuristic to
output decoded frames sooner.

This heuristic isn't perfect.  It's possible to make
an h264 stream that causes this heuristic to output
frames in an incorrect order, but not many frames,
and any potential out-of-order output is only near the
start of a stream.

We can disable this if/when the FW starts providing
max_num_reorder_frames.

For clear streams we could directly parse the SPS to
get max_num_reorder_frames, but the downsides of that
seem larger than the risk of this heuristic.

BUG: 58101
Change-Id: Id4ceb9de015c79274b69a0b3cf584f4450470f27
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/420976
Reviewed-by: John Bauman <jbauman@google.com>
Testability-Review: John Bauman <jbauman@google.com>
Commit-Queue: Dustin Green <dustingreen@google.com>