[time] Remove network_time integration test.
This test ensured the correct behavior of the original time architecture
where network time service was the coordinator, and won't work as we go
through the transition to the intended arhitecture where timekeeper is
the coordinator. Per the time test strategy we will eventually have a
timekeeper integration test offering similar coverage, but in the
interim we will rely on unittest and the time e2e test.

Change-Id: I1d282e0c600187b1d4dcf999b3503cb14544fe8b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/420721
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Testability-Review: Satsuki Ueno <satsukiu@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>