[power_manager][luis] Use ThermalShutdown node
The thermal policy on Luis is currently very simple and is only required
to shutdown the device at a configured temperature threshold.

The policy was initially set up to achieve this by using the existing
ThermalPolicy node. Now, we have the ThermalShutdown node which is a
dedicated node for this purpose.

We should use the ThermalShutdown node until we have a reason to use the
ThermalPolicy node on Luis. This will let us be more flexible in driver
transitions on Luis. Specifically, it lets us remove the CPU driver
(temporarily) on Luis without breaking power_manager dependencies.

Test: fx test power_manager_tests
Test: Boot up a Luis board with the change and verify the thermal
  shutdown logic behaves correctly by lowering the reboot temperature
  threshold and then artificially increasing the device temperature and
  observing a system reboot.
Change-Id: I8bd81210fd8111e8df711bc56fb9bc1485fd1782
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/420742
Reviewed-by: Justin Mattson <jmatt@google.com>
Reviewed-by: Onath Claridge <claridge@google.com>
Reviewed-by: Ruchira Ravoori <ravoorir@google.com>
Testability-Review: Onath Claridge <claridge@google.com>
Commit-Queue: Patrick Shickel <pshickel@google.com>