[rust][build] Add rustc_version_string as build arg
This allows forcing recompilation when using a custom build of the Rust
toolchain.

We attempt to do the right thing in all cases: Using the CIPD instance
ID when a prebuilt is used, and falling back to the build arg (defaulted
to empty string) otherwise.

The Zircon build always uses the value derived in the Fuchsia build.
However, for cases where the Zircon build is being run directly, we
duplicate the logic to find the CIPD instance ID.

Change-Id: I79613a9c035e83acb6b162115edf16f8146c2408
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/421735
Reviewed-by: Petr Hosek <phosek@google.com>
Reviewed-by: Didrik Nordstrom <dnordstrom@google.com>
Commit-Queue: Tyler Mandry <tmandry@google.com>