[wlan][policy] Wraps IfaceManager in service loop.
This change provides wlancfg's various components with a
request sender that is serviced by a loop that dispatches
requests to the existing IfaceManager implementations.  This
is in preparation for monitoring the lifetimes of client and
AP state machines in the policy layer and reacting to their
untimely completions.

Test: fx test wlancfg-tests
Multiply: wlancfg-tests
Bug: 57140
Change-Id: I1bcac3a3277982dfe5c2b6ef8f3e8e92042b753c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/418837
Reviewed-by: Marc Khouri <mnck@google.com>
Testability-Review: Marc Khouri <mnck@google.com>
Commit-Queue: Kevin Sakuma <sakuma@google.com>