[io][conformance] Make ulibfs a child component in conformance test
Deletes the io_fidl_conformance_tests package, which used to hardcode
the ulibfs test harness as the only harness it tested, and replaces it
with a new test package specifically for ulibfs. This construction can
be easily replicated for other filesystems in the future.

This also significantly simplifies the handshake between Io1Harness and
Io1HarnessReceiver. In fact, the receiver is not needed anymore at all,
so that has been deleted.

Bug: 33880
Change-Id: Ib5a27594d44994d784f9fa0f2a9905d8d98c2f71
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/422456
Reviewed-by: Chris Suter <csuter@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Testability-Review: Chris Suter <csuter@google.com>
Commit-Queue: Cameron Silvestrini <csilvestrini@google.com>