[fatfs][mounter] Fix race when opening block device
If we received two requests very close together, it's possible that both
would try and open the block device. Previously the code assumed that
both would succeed, and checked that the "current" block device was
still None before overwriting it.

Unfortunately, this assumption is incorrect and what would actually
happen is that one of the requests would get back Status::ALREADY_BOUND
(which occurs when the block FIFO is being bound). Fix this by switching
to an async-friendly mutex and holding that lock while initialising the
new FatDevice.

Bug: 59131
Change-Id: I8fbbb28cea9b851c483c8b8132340398ebc82647
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/422538
Commit-Queue: Simon Shields <simonshields@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Testability-Review: Chris Suter <csuter@google.com>