[camera][controller] Initial luis controller support
This CL provides initial controller support for luis.
It consists of a single configuration that provides a full-resolution
stream from the ISP.

The present form compiles and doesn't crash on a luis platform board.
Validation of the streaming is blocked on the recept of sensor frames.

Change-Id: Ifea9fcf3126180795b0c72560e28eb6087cd3706
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/422474
Commit-Queue: John Sasinowski <jsasinowski@google.com>
Reviewed-by: Nic Zoghb <nzo@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Testability-Review: Nic Zoghb <nzo@google.com>