[prebuilts] Update SwiftShader host build scripts.
(1) fxr/353570 was previously blocked by a bug in vulkan.hpp in
old Vulkan SDK. We update the SDK version to 1.2.148, and all
the dependency libraries.

(2) On some systems (e.g. Debian workstation) we saw a crash caused
by libvulkan.so.1 and libvulkan.so pointing to different files. Now
we make libvulkan.so a symbolic link to libvulkan.so.1 to avoid this.

(3) VkLayer_standard_validation is not supported by the SDK anymore.
We remove this from the build scripts.

TEST=./update-build-config.py > build_config
     ./build-linux-host-prebuilts.sh --install-dir=/tmp

Bug: 43647
Change-Id: I32399bf6974b71cbe859f94edce23bba9368e6ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/421746
Commit-Queue: Yilong Li <liyl@google.com>
Reviewed-by: David Turner <digit@google.com>
Testability-Review: Yilong Li <liyl@google.com>