[aml][cpu] Make CPU Bind rules board specific.
We want to use a different CPU for multiple boards that have
the same SoC and it's getting clunky to match on the correct
driver.

Make the drivers bind to board specific bind rules to get
around this.

Test: Booted on Sherlock/Astro, Device Enumeration
Change-Id: Ia839b1be1d02844826d571d9678a8f95cce2c2a9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/422478
Reviewed-by: Braden Kell <bradenkell@google.com>
Reviewed-by: Patrick Shickel <pshickel@google.com>
Testability-Review: Braden Kell <bradenkell@google.com>
Commit-Queue: Gurjant Kalsi <gkalsi@google.com>