[build] Consult custom_signing_script during flash.sh creation
Today we rely on a redundant `signed_image` variable to inform the
creation of the flashing script when to use signed images. (I hit a bug
in which we failed to set this for Sherlock.) This is the first of three
changes to migrate us off of `signed_image` and just consult
`custom_signing_script` directly.

Change-Id: I96f7261596b15f0058a90406e2dfce2e75a30516
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/423129
Reviewed-by: David Pursell <dpursell@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>