[camera][bin] Fix threading issue in FactoryServer
This CL fixes an issue whereby a callback was being passed into
RequestCapture on the factory_server thread but this was then mistakenly
invoked on the streamer thread. This lead to a crash on non-release
builds when trying to capture frames due to the `IsCurrentThreadOk()`
assert.

Test: capturing frames works correctly.
Fix: b/166159847

Change-Id: Ifbc4091bd0110965952aa23314ddbbad97e2eb76
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/422882
Reviewed-by: Matt Sandy <msandy@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Testability-Review: Matt Sandy <msandy@google.com>
Commit-Queue: Nic Zoghb <nzo@google.com>