[scenic] Fix racy annotation View crash
This CHECK will fire if the View-to-be-annotated's session is
destroyed before the AnnotationManager gets a chance to create the
annotation ViewHolder.  The dead session does not inform
AnnotationManager that the View is gone and the weak_ptr ends up
causing a CHECK.

The fix is simple: let AnnotationManager know that the View is already
gone so it can return ZX_ERR_PEER_CLOSED.  Since the Session doesn't have
access to the AnnotationManager, invalidate the entry inside of the
ViewTree to accomplish that feat indirectly.

TEST: Added tests to ViewTreeTest and AnnotationManagerTest
BUG: 58564
Change-Id: Ib1b98377c2d1602152815692a26340966b384ef7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/414462
Testability-Review: Yilong Li <liyl@google.com>
Reviewed-by: David Reveman <reveman@google.com>
Reviewed-by: Jaeheon Yi <jaeheon@google.com>
Reviewed-by: Yilong Li <liyl@google.com>
Commit-Queue: David Worsham <dworsham@google.com>