[zbitl] Add zbitl::MapUnownedVmo and zbitl::MapOwnedVmo
This is a special storage-view type to hold a zx::unowned_vmo along
with a zx::unowned_vmar that lets it use mapping rather than buffering
for payload access.  There's also a version that owns the VMO, which
is what Copy on the unowned type yields.

We add two test cases relevant to these types: multiple items within
page boundaries, and multiple items in which a subsequent item is on a
page boundary.

Bug: 57244
Change-Id: If99132bea61d9b561cad2768927a3dfaf0d5634e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/412746
Reviewed-by: George Kulakowski <kulakowski@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Testability-Review: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>